Should all be working now.

+++ sweep_log_streams.py
@@ -12,13 +12,13 @@ def print_log_group(log_group, message):
     print("[{}] {}".format(log_group['logGroupName'], message))
 
 
-def get_log_groups(prefix, nextToken=None):
+def get_log_groups(prefix, next_token=None):
     opts = {
         'logGroupNamePrefix': prefix,
         'limit': 50  # Maximum
     }
-    if nextToken:
-        opts['nextToken'] = nextToken
+    if next_token:
+        opts['nextToken'] = next_token
     log_groups_response = client.describe_log_groups(**opts)
     if log_groups_response:
         for log_group in log_groups_response['logGroups']:
@@ -28,13 +28,13 @@ def get_log_groups(prefix, nextToken=None):
             yield from get_log_groups(prefix, log_groups_response['nextToken'])
 
 
-def get_streams(log_group, nextToken=None):
+def get_streams(log_group, next_token=None):
     opts = {
         'logGroupName': log_group['logGroupName'],
         'limit': 50  # Max
     }
-    if nextToken:
-        opts['nextToken'] = nextToken
+    if next_token:
+        opts['nextToken'] = next_token
 
     response = client.describe_log_streams(**opts)
 
@@ -45,7 +45,7 @@ def get_streams(log_group, nextToken=None):
             yield from get_streams(log_group, response['nextToken'])
 
 
-def delete_old_streams(log_group):
+def delete_old_streams(log_group, dry_run=False):
     """
     Delete old log streams that are empty. Events get cleaned up by log_group['retentionInDays'] but the streams don't.
     """
@@ -71,13 +71,16 @@ def delete_old_streams(log_group):
             stream_time = datetime.fromtimestamp(stream['creationTime'] / 1000, tz=tz.tzutc())
 
         if stream_time < oldest_valid_event:
-            print_log_group(log_group, "Deleting stream: " + stream['logStreamName'])
-            delete_result = client.delete_delete_log_stream(
-                logGroupName=log_group['logGroupName'],
-                logStreamName=stream['logStreamName']
-            )
-            print(delete_result)
-            print_log_group(log_group, "Deleted stream: " + stream['logStreamName'])
+            if dry_run:
+                print_log_group(log_group, "Would delete stream: " + stream['logStreamName'] + " (--dry-run set)")
+            else:
+                print_log_group(log_group, "Deleting stream: " + stream['logStreamName'])
+                delete_result = client.delete_delete_log_stream(
+                    logGroupName=log_group['logGroupName'],
+                    logStreamName=stream['logStreamName']
+                )
+                print(delete_result)
+                print_log_group(log_group, "Deleted stream: " + stream['logStreamName'])
         else:
             print_log_group(log_group, "Checked stream, keeping: " + stream['logStreamName'])
 
@@ -85,18 +88,25 @@ def delete_old_streams(log_group):
 def get_arg_parser():
     parser = argparse.ArgumentParser(description="Cleans up old and empty log streams from log groups matching a "
                                                  "provided pattern")
+
+    parser.add_argument("--dry-run",
+                        dest="dry_run",
+                        action="store_true",
+                        help="Just print what we're going to do, don't actually do it."
+                        )
+
     parser.add_argument("prefix",
                         help="The log group prefix to filter for. Example: '/aws/lambda/app-staging-'"
                         )
     return parser
 
 
-def main(prefix):
+def main(prefix, dry_run=False):
     for log_group in get_log_groups(prefix):
-        delete_old_streams(log_group)
+        delete_old_streams(log_group, dry_run)
     print("Done")
 
 
 if __name__ == "__main__":
     args = get_arg_parser().parse_args()
-    main(args.prefix)
+    main(args.prefix, args.dry_run)
